Fix cascade bug solves #foundations/06-cascade-fix
The order of css rules of same specificity in `style.css` caused the
cascade to overwrite styles applied earlier in the file.
Applies to rule pairs:
    - .para -> .small-para
    - .button -> .confirm

A similar issue occurs with the classes:
    - div.text -> .child
However reordering those rules still left the former with higher
specificity, requiring further intervention with explicitly increasing
the specificity.